search service decomposed but needs to remake
